create an array

+++ malloc_free/0-create_array.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include "main.h"
 
+
 /**
  * create_array - create array of chars and initialize with a specific char
  * @size: size of array
@@ -22,7 +23,7 @@ char *create_array(unsigned int size, char c)
 	if (tab == NULL)
 		return (NULL);
 
-	for (i = 0; i < size ; i++)
+	for (i = 0; i < (int)size ; i++)
 	{
 		*(tab + i) = c;
 	}
